move ChannelsLastHparams
