+++ composer/algorithms/algorithm_registry.py
@@ -3,10 +3,10 @@
 from typing import Dict, List, Type
 
 from composer.algorithms.algorithm_hparams import AlgorithmHparams
-from composer.algorithms.channels_last import ChannelsLastHparams
-from composer.algorithms.hparams import (AlibiHparams, AugMixHparams, BlurPoolHparams, ColOutHparams, CutMixHparams,
-                                         CutOutHparams, FactorizeHparams, GhostBatchNormHparams, LabelSmoothingHparams,
-                                         MixUpHparams, RandAugmentHparams, SqueezeExciteHparams, StochasticDepthHparams)
+from composer.algorithms.hparams import (AlibiHparams, AugMixHparams, BlurPoolHparams, ChannelsLastHparams,
+                                         ColOutHparams, CutMixHparams, CutOutHparams, FactorizeHparams,
+                                         GhostBatchNormHparams, LabelSmoothingHparams, MixUpHparams, RandAugmentHparams,
+                                         SqueezeExciteHparams, StochasticDepthHparams)
 from composer.algorithms.layer_freezing import LayerFreezingHparams
 from composer.algorithms.no_op_model import NoOpModelHparams
 from composer.algorithms.progressive_resizing import ProgressiveResizingHparams
